requirements.txt
# Core
python-dotenv==1.0.1
requests==2.31.0
httpx==0.27.0

# Apify SDK
apify-client==1.6.4

# Google APIs
google-generativeai==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0

# Data processing
pandas==2.2.1

# Logging and utilities
loguru==0.7.2
pydantic==2.6.4
pydantic-settings==2.2.1

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6

# Development
black==24.3.0
ruff==0.3.4
